# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API RESTful para Gerenciamento de Tarefas (To-Do List) com Laravel",
        "description": "Documentação da API de gerenciamento de tarefas.",
        "contact": {
            "email": "jeandcabreu@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/tasks": {
            "get": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Lista todas as tarefas com seus status",
                "description": "Retorna uma lista de todas as tarefas cadastradas, incluindo os status associados.",
                "operationId": "getTasks",
                "responses": {
                    "200": {
                        "description": "Lista de tarefas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar tarefas: Mensagem do erro..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Criar uma nova tarefa",
                "operationId": "4dbf3e760f46296529098c64a16b6549",
                "requestBody": {
                    "description": "Dados da tarefa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "status_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Criar Modal"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Modal para exibir detalhes da tarefa"
                                    },
                                    "status_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarefa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{hash}/status": {
            "patch": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Atualiza o status de uma tarefa",
                "description": "Recebe um hash e um status_id para atualizar o status da tarefa correspondente.",
                "operationId": "updateTaskStatus",
                "parameters": [
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Hash único da tarefa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123def456"
                        }
                    }
                ],
                "requestBody": {
                    "description": "ID do novo status",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_id"
                                ],
                                "properties": {
                                    "status_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Status atualizado com sucesso"
                                        },
                                        "task": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação ou processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao atualizar status: Mensagem do erro..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{hash}": {
            "delete": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Remove uma tarefa pelo hash",
                "description": "Remove permanentemente uma tarefa específica identificada pelo seu hash único.",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Hash único da tarefa a ser removida",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123def456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa removida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa deletada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarefa não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa não encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao deletar tarefa: Mensagem do erro..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/status/{status}": {
            "get": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Lista tarefas por status",
                "description": "Retorna todas as tarefas filtradas por um status específico (pendente, em-andamento, concluida).",
                "operationId": "getTasksByStatus",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Status das tarefas a serem filtradas",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pendente",
                                "em-andamento",
                                "concluida"
                            ],
                            "example": "pendente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tarefas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Status inválido ou erro no processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Status inválido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Status não encontrado no banco de dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar tarefas por status: Mensagem do erro..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Tarefas",
            "description": "Tarefas"
        }
    ]
}